# Move `log.Processor.Enabled` to independent `FilterProcessor` interfaced type