--- a/sdk/log/simple_test.go
+++ b/sdk/log/simple_test.go
@@ -50,12 +50,6 @@ func TestSimpleProcessorOnEmit(t *testing.T) {
 	assert.Equal(t, []log.Record{*r}, e.records)
 }
 
-func TestSimpleProcessorEnabled(t *testing.T) {
-	e := new(exporter)
-	s := log.NewSimpleProcessor(e)
-	assert.True(t, s.Enabled(context.Background(), log.Record{}))
-}
-
 func TestSimpleProcessorShutdown(t *testing.T) {
 	e := new(exporter)
 	s := log.NewSimpleProcessor(e)
@@ -76,7 +70,6 @@ func TestSimpleProcessorEmpty(t *testing.T) {
 		ctx := context.Background()
 		record := new(log.Record)
 		assert.NoError(t, s.OnEmit(ctx, record), "OnEmit")
-		assert.False(t, s.Enabled(ctx, *record), "Enabled")
 		assert.NoError(t, s.ForceFlush(ctx), "ForceFlush")
 		assert.NoError(t, s.Shutdown(ctx), "Shutdown")
 	})
@@ -97,7 +90,6 @@ func TestSimpleProcessorConcurrentSafe(t *testing.T) {
 			defer wg.Done()
 
 			_ = s.OnEmit(ctx, r)
-			_ = s.Enabled(ctx, *r)
 			_ = s.Shutdown(ctx)
 			_ = s.ForceFlush(ctx)
 		}()