@pellared 2024-08-08
> The `Processor` interface is now compliant with the OTel spec 

Is adding additional functionalities against the specification?

Notice that `WithProcessor` right now also handles `FilteredProcessor` which means that our SDK handles processors with `Enabled` method.  Basically it is a kind of "implicit" interface that is optional for users to implement. rom specification perspective, I do not see any difference because we still support for `Enabled` for processors added to the `LoggerProvider`. When introducing `FilteredProcessor` we would keep the backward compatibility of the SDK.

I think the introduced pattern is good for maintaining backwards compatibility (when we **already** have stable interfaces) or to add support optional behavior that the users does not need to explicitly implement. However, using this pattern makes "decorators" more error-prone (e.g. see that the example is more complex).