@MrAlias 2024-08-08
> Is adding additional functionalities against the specification?

No, it is not[^1].

[^1]: https://github.com/open-telemetry/opentelemetry-specification/tree/main/specification#notation-conventions-and-compliance