@MrAlias 2024-08-08
> I think we can should `Enabled` from `SimpleProcessor` and `BatchProcessor`.

I don't follow. Both implementations still implement `FilterProcessor`?