@MrAlias 2024-08-08
> However, using this pattern makes "decorators" more error-prone (e.g. see that the example is more complex).

Can you explain how this pattern is more error prone? I don't follow. Is there an error we need to address?