@MrAlias 2024-08-08
> I think the proposed pattern is good for maintaining backwards compatibility (when we **already** have stable interfaces) or to add support optional behavior that the users does not need to explicitly implement.

Isn't this also a good pattern when we do not want all Processors to have to implement this method and we want to extend our implementations functionality beyond the stable specification definition of what a Processor is?