@pellared 2024-08-08
> > Is adding additional functionalities against the specification?
> 
> No, it is not[1](#user-content-fn-1-e63acccb31954194cead1d7de5d3a769).
> 
> ## Footnotes
> 1. https://github.com/open-telemetry/opentelemetry-specification/tree/main/specification#notation-conventions-and-compliance [↩](#user-content-fnref-1-e63acccb31954194cead1d7de5d3a769)



> Isn't this also a good pattern when we do not want all Processors to have to implement this method and we want to extend our implementations functionality beyond the stable specification definition of what a Processor interface is?