@pellared 2024-08-08
> Isn't this also a good pattern when we do not want all Processors to have to implement this method

If so then I think we should consider doing the same for `ForceFlush` and `Shutdown`.