@pellared 2024-08-08
> > However, using this pattern makes "decorators" more error-prone (e.g. see that the example is more complex).
> 
> Can you explain how this pattern is more error prone? I don't follow. Is there an error we need to address?

I mean

```go
func (p *ContextFilterProcessor) Enabled(ctx context.Context, record log.Record) bool {
	p.lazyFilter.Do(func() {
		if f, ok := p.Processor.(log.FilterProcessor); ok {
			p.filter = f
		}
	})
	return !ignoreLogs(ctx) && (p.filter == nil || p.filter.Enabled(ctx, record))
}
```

Each processor decorator would need to implement all possible methods that the SDK uses.

PS. This is not a blocking concern.