@MrAlias 2024-08-08
> I do not understand why adding explicitly `Enabled` to Processor would violate the specification whereas adding `Enabled` implicitly via `FilterProcessor` would not violate the specification. For me these are only "language specifics".

Hmm,  I didn't mean to imply the existing code violates the specification. It would not be a specification incomparable change to add `Enabled` to the `Processor` interface. It is more of a concern that the specification would evolve to add that to the interface in an incompatible way to what we implement.

> From user perspective we are still adding `Enabled` to processors.

Right, I think this is the important point we are trying address. The language specifics of Java and other languages mean that a user there would not care about this type of functionality. Here in Go this can be an important feature to add.