@MrAlias 2024-08-08
> If so then I think we should consider doing the same for `ForceFlush` and `Shutdown`. However, this would be not similar to the design of Trace and Metrics SDKs.

Yeah, I think that is all correct. I would like to have split these up, but not only prior signal precedent inhibits that, but also the specification explicitly.