@MrAlias 2024-08-08
> Each processor decorator would need to implement all possible methods that the SDK uses.

That does not seem correct. This is explicitly designed to be the opposite of what you say.