@pellared 2024-08-08
> That does not seem correct. This is explicitly designed to be the opposite of what you say.

I mean only  **decorators** which wrap other processor(s).