@pellared 2024-08-08
> It is more of a concern that the specification would evolve to add that to the interface in an incompatible way to what we implement.

There would be still a problem if an `Enabled` would be added to the specification but with different signature. We would need to name the new method differently than in the specification.