@MrAlias 2024-08-08
SIG meeting notes:

- Do not export FilterProcessor
- Document the additional functionality (i.e. providing an `Enabled` method) as supported but experimental features
- Remove the `Enabled` methods from the batch and simple processors